test(util-stream): switch lambda test from e2e to integration
