--- packages/util-stream/jest.config.e2e.js
@@ -1,4 +1,0 @@
-module.exports = {
-  preset: "ts-jest",
-  testMatch: ["**/*.e2e.spec.ts"],
-};
