+++ packages/util-stream/package.json
@@ -10,8 +10,7 @@
     "build:types:downlevel": "downlevel-dts dist-types dist-types/ts3.4",
     "clean": "rimraf ./dist-* && rimraf *.tsbuildinfo",
     "test": "jest && karma start karma.conf.js",
-    "test:integration": "jest -c jest.config.integ.js",
-    "test:e2e": "jest -c jest.config.e2e.js"
+    "test:integration": "jest -c jest.config.integ.js"
   },
   "main": "./dist-cjs/index.js",
   "module": "./dist-es/index.js",
