--- packages/util-stream/test/index.mjs
@@ -1,6 +1,0 @@
-/**
- * function.zip contains this file.
- */
-export const handler = async (event) => {
-  return event;
-};
